+++ assets/scripts/settings.js
@@ -1,10 +1,46 @@
 document.addEventListener("DOMContentLoaded", ()=> {
     const baseUrl = BE_URL;
+    let originalData = {};
 
-    logoutBtn.addEventListener("click", () => {
-        localStorage.clear();
-        window.location.href = "../index.html";
-    });
+    function loadComponent(id, file) {
+        fetch(file)
+            .then(response => response.text())
+            .then(data => {
+                document.getElementById(id).innerHTML = data;
+                attachNavbarEventListeners();
+            });
+    }
+    
+    function attachNavbarEventListeners() {
+        const logoutBtn = document.getElementById("logoutBtn");
+        if (logoutBtn) {
+            logoutBtn.addEventListener("click", () => {
+                localStorage.clear();
+                window.location.href = "../index.html";
+            });
+        }
+    }
+    
+    function highlightActiveLink() {
+        document.getElementById("current-year").textContent = new Date().getFullYear();
+        const footer = document.querySelector("footer");
+        function checkScrollbar() {
+            if (document.body.scrollHeight <= window.innerHeight) {
+                footer.classList.add("fixed");
+            } else {
+                footer.classList.remove("fixed");
+            }
+        }
+        checkScrollbar();
+        window.addEventListener("resize", checkScrollbar);
+        const currentPath = window.location.pathname;
+        const navLinks = document.querySelectorAll("nav ul li a");
+        navLinks.forEach(link => {
+            if (currentPath.endsWith(link.getAttribute("href"))) {
+                link.classList.add("active");
+            }
+        });
+    }
 
     document.getElementById("show-password").addEventListener("click", function () {
         var passwordFields = document.querySelectorAll(".password");
@@ -65,4 +101,30 @@ document.addEventListener("DOMContentLoaded", ()=> {
             alert(errorMessage);
         }
     }
+    
+    function getTeachersData() {
+        const token = localStorage.getItem("access_token");
+        fetch(`${baseUrl}/getSpecificTeacher/`, {
+            method: "GET",
+            headers: {
+                Authorization: `Bearer ${token}`,
+                "Content-Type": "application/json",
+            },
+        })
+        .then(handleResponse)
+        .then(data => {
+            originalData = { ...data };
+            if(originalData.teacher_type==="hod") {
+                loadComponent("navbar-hod", "../components/hod_navbar.html");
+            }
+            else {
+                loadComponent("navbar-faculty", "../components/faculty_navbar.html");
+            }
+            loadComponent("footer", "../components/footer.html");
+            setTimeout(highlightActiveLink, 100);
+        })
+        .catch(showError);
+    }
+
+    getTeachersData();
 });
